Host side endpoint refactored separately
