+++ src/main/java/com/teragrep/cfe18/handlers/HostRelpController.java
@@ -1,0 +1,212 @@
+/*
+ * Integration main data management for Teragrep
+ * Copyright (C) 2021  Suomen Kanuuna Oy
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://github.com/teragrep/teragrep/blob/main/LICENSE>.
+ *
+ *
+ * Additional permission under GNU Affero General Public License version 3
+ * section 7
+ *
+ * If you modify this Program, or any covered work, by linking or combining it
+ * with other code, such other code is not for that reason alone subject to any
+ * of the requirements of the GNU Affero GPL version 3 as long as this Program
+ * is the same Program as licensed from Suomen Kanuuna Oy without any additional
+ * modifications.
+ *
+ * Supplemented terms under GNU Affero General Public License version 3
+ * section 7
+ *
+ * Origin of the software must be attributed to Suomen Kanuuna Oy. Any modified
+ * versions must be marked as "Modified version of" The Program.
+ *
+ * Names of the licensors and authors may not be used for publicity purposes.
+ *
+ * No rights are granted for use of trade names, trademarks, or service marks
+ * which are in The Program if any.
+ *
+ * Licensee must indemnify licensors and authors for any liability that these
+ * contractual assumptions impose on licensors and authors.
+ *
+ * To the extent this program is licensed as part of the Commercial versions of
+ * Teragrep, the applicable Commercial License may apply to this file if you as
+ * a licensee so wish it.
+ */
+package com.teragrep.cfe18.handlers;
+
+import com.teragrep.cfe18.HostRelpMapper;
+import com.teragrep.cfe18.handlers.entities.HostFile;
+import com.teragrep.cfe18.handlers.entities.HostRelp;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
+import org.json.JSONObject;
+import org.mybatis.spring.SqlSessionTemplate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import javax.sql.DataSource;
+import java.sql.SQLException;
+import java.util.List;
+
+@RestController
+@RequestMapping(path = "host/relp")
+@SecurityRequirement(name = "api")
+public class HostRelpController {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(HostRelpController.class);
+
+    @Autowired
+    DataSource dataSource;
+
+    @Autowired
+    SqlSessionTemplate sqlSessionTemplate;
+
+    @Autowired
+    HostRelpMapper hostRelpMapper;
+
+
+    @RequestMapping(path = "", method = RequestMethod.PUT, produces = "application/json")
+    @Operation(summary = "Create relp based host")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "New relp based host created",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = HostRelp.class))}),
+            @ApiResponse(responseCode = "400", description = "ID,MD5 or fqhost already exists",
+                    content = @Content),
+            @ApiResponse(responseCode = "400", description = "Host exists with different type",
+                    content = @Content),
+            @ApiResponse(responseCode = "500", description = "Internal server error, contact admin", content = @Content)})
+    public ResponseEntity<String> create(@RequestBody HostRelp newHostRelp) {
+        LOGGER.info("About to insert <[{}]>", newHostRelp);
+        try {
+            HostRelp hr = hostRelpMapper.create(
+                    newHostRelp.getMd5(),
+                    newHostRelp.getFqHost());
+            LOGGER.debug("Values returned <[{}]>", hr);
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put("id", hr.getId());
+            jsonObject.put("message", "New host created with relp type");
+            return new ResponseEntity<>(jsonObject.toString(), HttpStatus.CREATED);
+        } catch (RuntimeException ex) {
+            JSONObject jsonErr = new JSONObject();
+            jsonErr.put("id", newHostRelp.getId());
+            jsonErr.put("message", ex.getCause().getMessage());
+            final Throwable cause = ex.getCause();
+            if (cause instanceof SQLException) {
+                LOGGER.error((cause).getMessage());
+                // Get specific error type
+                int error = ((SQLException) cause).getErrorCode();
+                // Link error with state to get accurate error status
+                String state = error + "-" + ((SQLException) cause).getSQLState();
+                if (state.equals("1062-23000")) {
+                    jsonErr.put("message", "ID,MD5 or fqhost already exists");
+                    return new ResponseEntity<>(jsonErr.toString(), HttpStatus.BAD_REQUEST);
+                }else if (state.equals("45000")) {
+                    jsonErr.put("message", "Host exists with different type");
+                    return new ResponseEntity<>(jsonErr.toString(), HttpStatus.BAD_REQUEST);
+                }
+            }
+            return new ResponseEntity<>(jsonErr.toString(), HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @RequestMapping(path = "/{id}", method = RequestMethod.GET, produces = "application/json")
+    @Operation(summary = "Fetch relp based host by ID")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Relp based host retrieved",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = HostRelp.class))}),
+            @ApiResponse(responseCode = "400", description = "Host does not exist",
+                    content = @Content),
+            @ApiResponse(responseCode = "500", description = "Internal server error, contact admin", content = @Content)})
+    public ResponseEntity<?> get(@PathVariable("id") int id, @RequestParam(required = false) Integer version) {
+        try {
+            HostRelp hr = hostRelpMapper.get(id,version);
+            return new ResponseEntity<>(hr, HttpStatus.OK);
+        } catch (Exception ex) {
+            JSONObject jsonErr = new JSONObject();
+            jsonErr.put("id", id);
+            jsonErr.put("message", ex.getCause().getMessage());
+            final Throwable cause = ex.getCause();
+            if (cause instanceof SQLException) {
+                LOGGER.error((cause).getMessage());
+                String state = ((SQLException) cause).getSQLState();
+                if (state.equals("45000")) {
+                    jsonErr.put("message", "Record does not exist");
+                    return new ResponseEntity<>(jsonErr.toString(), HttpStatus.NOT_FOUND);
+                }
+            }
+            return new ResponseEntity<>(jsonErr.toString(), HttpStatus.BAD_REQUEST);
+        }
+    }
+
+    @RequestMapping(path = "", method = RequestMethod.GET, produces = "application/json")
+    @Operation(summary = "Fetch all relp hosts", description = "Will return empty list if there are no hosts to fetch")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = HostRelp.class))})})
+    public List<HostRelp> getAll(@RequestParam(required = false) Integer version) {
+        return hostRelpMapper.getAll(version);
+    }
+
+    @RequestMapping(path = "/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
+    @Operation(summary = "Delete relp host")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Host deleted",
+                    content = {@Content(mediaType = "application/json",
+                            schema = @Schema(implementation = HostRelp.class))}),
+            @ApiResponse(responseCode = "400", description = "Host is being used OR Host does not exist",
+                    content = @Content),
+            @ApiResponse(responseCode = "500", description = "Internal server error, contact admin", content = @Content)
+    })
+    public ResponseEntity<String> delete(@PathVariable("id") int id) {
+        LOGGER.info("Deleting Host  <[{}]>",id);
+        try {
+            hostRelpMapper.delete(id);
+            JSONObject j = new JSONObject();
+            j.put("id", id);
+            j.put("message", "Host deleted");
+            return new ResponseEntity<>(j.toString(), HttpStatus.OK);
+        } catch (RuntimeException ex) {
+            JSONObject jsonErr = new JSONObject();
+            jsonErr.put("id", id);
+            jsonErr.put("message", ex.getCause().getMessage());
+            final Throwable cause = ex.getCause();
+            if (cause instanceof SQLException) {
+                LOGGER.error((cause).getMessage());
+                String state = ((SQLException) cause).getSQLState();
+                if (state.equals("23000")) {
+                    jsonErr.put("message", "Is in use");
+                    return new ResponseEntity<>(jsonErr.toString(), HttpStatus.BAD_REQUEST);
+                } else if (state.equals("45000")) {
+                    jsonErr.put("message", "Record does not exist");
+                    return new ResponseEntity<>(jsonErr.toString(), HttpStatus.NOT_FOUND);
+                }
+            }
+            return new ResponseEntity<>(jsonErr.toString(), HttpStatus.BAD_REQUEST);
+        }
+    }
+}
+
+
