--- src/main/resources/com/teragrep/cfe18/HostMapper.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.teragrep.cfe18.HostMapper">
-
-    <!-- Get host file by ID-->
-    <select id="getHostFileById" resultMap="HostFile" statementType="CALLABLE">
-        {CALL cfe_00.retrieve_host_details(#{param1},#{param2})}
-    </select>
-    <resultMap type="com.teragrep.cfe18.handlers.entities.HostFile" id="HostFile">
-        <result property="id" column="host_id"/>
-        <result property="md5" column="host_md5"/>
-        <result property="fqHost" column="host_fq"/>
-        <result property="host_type" column="host_type"/>
-        <result property="hub" column="hub_id"/>
-        <result property="hostname" column="host_name"/>
-        <result property="host_meta_id" column="host_meta_id"/>
-        <result property="hub_fq" column="hub_fq"/>
-    </resultMap>
-
-
-    <!-- Insert new host file -->
-    <select id="addHostFile" resultMap="insertedID" statementType="CALLABLE">
-        {CALL
-        location.host_add_cfe(#{param1},#{param2},#{param3})}
-    </select>
-    <resultMap id="insertedID" type="com.teragrep.cfe18.handlers.entities.HostFile">
-        <result property="id" column="last"/>
-    </resultMap>
-
-
-    <!-- Get relp host by ID-->
-    <select id="getHostRelpById" resultMap="HostRelp" statementType="CALLABLE">
-        {CALL cfe_00.retrieve_host_details(#{param1},#{param2})}
-    </select>
-    <resultMap type="com.teragrep.cfe18.handlers.entities.HostRelp" id="HostRelp">
-        <result property="id" column="host_id"/>
-        <result property="md5" column="md5"/>
-        <result property="fqHost" column="fqhost"/>
-        <result property="hostType" column="host_type"/>
-    </resultMap>
-
-
-    <!-- Insert new relp host -->
-    <select id="addHostRelp" resultMap="insertedIDRelp" statementType="CALLABLE">
-        {CALL
-        location.host_add_relp(#{param1},#{param2})}
-    </select>
-    <resultMap id="insertedIDRelp" type="com.teragrep.cfe18.handlers.entities.HostRelp">
-        <result property="id" column="last"/>
-    </resultMap>
-
-
-    <!-- GET ALL Host -->
-    <select id="getAllHost" resultMap="allHost" statementType="CALLABLE">
-        {CALL location.retrieve_all_hosts(#{param1})}
-    </select>
-    <resultMap id="allHost" type="com.teragrep.cfe18.handlers.entities.HostFile">
-        <result property="id" column="host_id"/>
-        <result property="fqHost" column="host_fq"/>
-        <result property="md5" column="host_md5"/>
-        <result property="host_type" column="host_type"/>
-        <result property="hostname" column="hostname"/>
-        <result property="host_meta_id" column="host_meta_id"/>
-        <result property="hub_fq" column="hub_fq"/>
-        <result property="hub" column="hub_id"/>
-    </resultMap>
-
-
-    <!-- Delete -->
-    <select id="deleteHost" statementType="CALLABLE">
-        {CALL location.remove_host(#{param1})}
-    </select>
-
-</mapper>
-
