+++ src/main/resources/com/teragrep/cfe18/HostRelpMapper.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.teragrep.cfe18.HostRelpMapper">
+
+    <!-- Insert relp  ( proc_MD5 VARCHAR(32), proc_fqhost VARCHAR(128) ) -->
+    <select id="create" resultMap="insertedID" statementType="CALLABLE">
+        {CALL
+        location.insert_relp_host(#{param1},#{param2})}
+    </select>
+
+
+    <!-- get    ( proc_host_id INT, tx_id INT ) -->
+    <select id="get" resultMap="getRelpHost" statementType="CALLABLE">
+        {CALL cfe_00.select_relp_host(#{param1},#{param2})}
+    </select>
+    <resultMap id="getRelpHost" type="com.teragrep.cfe18.handlers.entities.HostRelp">
+        <result property="id"               column="id"/>
+        <result property="md5"              column="host_md5"/>
+        <result property="fqHost"           column="host_fq"/>
+    </resultMap>
+
+
+    <!-- GET ALL Relp  ( tx_id INT ) -->
+    <select id="getAll" resultMap="allRelp" statementType="CALLABLE">
+        {CALL location.select_all_relp_hosts(#{param1})}
+    </select>
+    <resultMap id="allRelp" type="com.teragrep.cfe18.handlers.entities.HostRelp">
+        <result property="id"               column="id"/>
+        <result property="md5"              column="host_md5"/>
+        <result property="fqHost"           column="host_fq"/>
+    </resultMap>
+
+
+    <!-- Delete Relp  ( proc_host_id INT ) -->
+    <select id="delete" statementType="CALLABLE">
+        {CALL location.delete_relp_host(#{param1})}
+    </select>
+
+
+    <!-- ID return map -->
+    <resultMap id="insertedID" type="com.teragrep.cfe18.handlers.entities.HostRelp">
+        <result property="id"       column="id"/>
+    </resultMap>
+
+</mapper>
+
