+++ src/main/resources/db/migration/R__Procedure_Retrieve_Host_File.sql
@@ -43,48 +43,46 @@
  * Teragrep, the applicable Commercial License may apply to this file if you as
  * a licensee so wish it.
  */
-package com.teragrep.cfe18.handlers.entities;
+USE cfe_00;
+DELIMITER //
+CREATE OR REPLACE PROCEDURE select_file_host(proc_host_id INT, tx_id INT)
+BEGIN
+    DECLARE EXIT HANDLER FOR SQLEXCEPTION
+        BEGIN
+            ROLLBACK;
+            RESIGNAL;
+        END;
+    START TRANSACTION;
+    IF (tx_id) IS NULL THEN
+        SET @time = (SELECT MAX(transaction_id) FROM mysql.transaction_registry);
+    ELSE
+        SET @time = tx_id;
+    END IF;
+    IF ((SELECT COUNT(id) FROM location.host FOR SYSTEM_TIME AS OF TRANSACTION @time WHERE id = proc_host_id) = 0) THEN
+        SELECT JSON_OBJECT('id', proc_host_id, 'message', 'Host does not exist') INTO @hid;
+        SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = @hid;
+    END IF;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import io.swagger.v3.oas.annotations.media.Schema;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class HostRelp {
-    @Schema(accessMode = Schema.AccessMode.READ_ONLY)
-    private int id;
-    private String md5;
-    private String fqHost;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getMd5() {
-        return md5;
-    }
-
-    public void setMd5(String md5) {
-        this.md5 = md5;
-    }
-
-    public String getFqHost() {
-        return fqHost;
-    }
-
-    public void setFqHost(String fqHost) {
-        this.fqHost = fqHost;
-    }
-
-    @Override
-    public String toString() {
-        return "HostRelp{" +
-                "id=" + id +
-                ", md5='" + md5 + '\'' +
-                ", fqHost='" + fqHost + '\'' +
-                '}';
-    }
-}
+    IF ((SELECT COUNT(h.id)
+         FROM location.host FOR SYSTEM_TIME AS OF TRANSACTION @time h
+                  INNER JOIN hubs FOR SYSTEM_TIME AS OF TRANSACTION @time h3 ON h.id = h3.host_id
+         WHERE h.id = proc_host_id) > 0) THEN
+        SELECT JSON_OBJECT('id', proc_host_id, 'message', 'Host is hub') INTO @hub;
+        SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = @hub;
+    ELSE
+        SELECT h.id          AS id,
+               h.md5         AS host_md5,
+               h.fqhost      AS host_fq,
+               h2.id         AS hub_id,
+               h3.fqhost     AS hub_fq
+        FROM location.host FOR SYSTEM_TIME AS OF TRANSACTION @time h
+                 INNER JOIN host_type_cfe FOR SYSTEM_TIME AS OF TRANSACTION @time htc ON h.id = htc.host_id
+                 INNER JOIN hubs FOR SYSTEM_TIME AS OF TRANSACTION @time h2 ON htc.hub_id = h2.id
+                 INNER JOIN location.host FOR SYSTEM_TIME AS OF TRANSACTION @time h3 ON h2.host_id = h3.id
+        WHERE h.id = proc_host_id
+          AND h3.id = h2.host_id;
+    END IF;
+    COMMIT;
+END;
+//
+DELIMITER ;
