+++ src/test/java/com/teragrep/cfe18/controllerTests/HostFileControllerTest.java
@@ -1,0 +1,301 @@
+/*
+ * Integration main data management for Teragrep
+ * Copyright (C) 2021  Suomen Kanuuna Oy
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://github.com/teragrep/teragrep/blob/main/LICENSE>.
+ *
+ *
+ * Additional permission under GNU Affero General Public License version 3
+ * section 7
+ *
+ * If you modify this Program, or any covered work, by linking or combining it
+ * with other code, such other code is not for that reason alone subject to any
+ * of the requirements of the GNU Affero GPL version 3 as long as this Program
+ * is the same Program as licensed from Suomen Kanuuna Oy without any additional
+ * modifications.
+ *
+ * Supplemented terms under GNU Affero General Public License version 3
+ * section 7
+ *
+ * Origin of the software must be attributed to Suomen Kanuuna Oy. Any modified
+ * versions must be marked as "Modified version of" The Program.
+ *
+ * Names of the licensors and authors may not be used for publicity purposes.
+ *
+ * No rights are granted for use of trade names, trademarks, or service marks
+ * which are in The Program if any.
+ *
+ * Licensee must indemnify licensors and authors for any liability that these
+ * contractual assumptions impose on licensors and authors.
+ *
+ * To the extent this program is licensed as part of the Commercial versions of
+ * Teragrep, the applicable Commercial License may apply to this file if you as
+ * a licensee so wish it.
+ */
+package com.teragrep.cfe18.controllerTests;
+
+import com.google.gson.Gson;
+import com.teragrep.cfe18.handlers.entities.HostFile;
+import com.teragrep.cfe18.handlers.entities.Hub;
+import org.apache.http.HttpEntity;
+import org.apache.http.HttpResponse;
+import org.apache.http.HttpStatus;
+import org.apache.http.client.methods.HttpDelete;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.client.methods.HttpPut;
+import org.apache.http.entity.ContentType;
+import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.HttpClientBuilder;
+import org.apache.http.util.EntityUtils;
+import org.json.JSONObject;
+import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+import java.util.ArrayList;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
+@ExtendWith(MigrateDatabaseExtension.class)
+public class HostFileControllerTest extends TestSpringBootInformation {
+
+    Gson gson = new Gson();
+
+    @LocalServerPort
+    private int port;
+
+    @Test
+    @BeforeAll
+    public void testData() throws Exception {
+        // insert hub first
+        // send one hub
+        Hub hub1 = new Hub();
+        hub1.setFqHost("hubfq");
+        hub1.setMd5("hubmd5");
+        hub1.setIp("hubip");
+
+        String json1 = gson.toJson(hub1);
+
+        // forms the json to requestEntity
+        StringEntity requestEntity1 = new StringEntity(
+                String.valueOf(json1),
+                ContentType.APPLICATION_JSON);
+
+        // Creates the request
+        HttpPut request1 = new HttpPut("http://localhost:" + port + "/host/hub");
+        // set requestEntity to the put request
+        request1.setEntity(requestEntity1);
+        // Header
+        request1.setHeader("Authorization", "Bearer " + token);
+
+        // Get the response from endpoint
+        HttpResponse httpResponse = HttpClientBuilder.create().build().execute(request1);
+
+        // Get the entity from response
+        HttpEntity entity = httpResponse.getEntity();
+
+        // Entity response string
+        String responseString = EntityUtils.toString(entity);
+
+        // Parsin respponse as JSONObject
+        JSONObject responseAsJson = new JSONObject(responseString);
+
+        // Creating expected message as JSON Object from the data that was sent towards endpoint
+        String expected = "New hub created";
+
+        // Creating string from Json that was given as a response
+        String actual = responseAsJson.get("message").toString();
+
+        // Assertions
+        assertEquals(expected, actual);
+        assertThat(
+                httpResponse.getStatusLine().getStatusCode(),
+                equalTo(HttpStatus.SC_CREATED));
+    }
+
+    @Test
+    @Order(1)
+    public void testInsertFileHost() throws Exception {
+        HostFile host = new HostFile();
+        host.setMd5("randommd5value");
+        host.setFqHost("hostFq");
+        host.setHubFq("hubfq");
+
+        String json = gson.toJson(host);
+
+        // forms the json to requestEntity
+        StringEntity requestEntity = new StringEntity(
+                String.valueOf(json),
+                ContentType.APPLICATION_JSON);
+
+        // Creates the request
+        HttpPut request = new HttpPut("http://localhost:" + port + "/host/file");
+        // set requestEntity to the put request
+        request.setEntity(requestEntity);
+        // Header
+        request.setHeader("Authorization", "Bearer " + token);
+
+        // Get the response from endpoint
+        HttpResponse httpResponse = HttpClientBuilder.create().build().execute(request);
+
+        // Get the entity from response
+        HttpEntity entity = httpResponse.getEntity();
+
+        // Entity response string
+        String responseString = EntityUtils.toString(entity);
+
+        // Parsin respponse as JSONObject
+        JSONObject responseAsJson = new JSONObject(responseString);
+
+        // Creating expected message as JSON Object from the data that was sent towards endpoint
+        String expected = "New host created";
+
+        // Creating string from Json that was given as a response
+        String actual = responseAsJson.get("message").toString();
+
+        // Assertions
+        assertEquals(expected, actual);
+        assertThat(
+                httpResponse.getStatusLine().getStatusCode(),
+                equalTo(HttpStatus.SC_CREATED));
+
+    }
+
+    @Test
+    @Order(2)
+    public void testGetFileHost() throws Exception {
+
+        HostFile host = new HostFile();
+        host.setId(2);
+        host.setMd5("randommd5value");
+        host.setFqHost("hostFq");
+        host.setHubId(1);
+        host.setHubFq("hubfq");
+
+        String json = gson.toJson(host);
+
+        // Asserting get request
+        HttpGet requestGet = new HttpGet("http://localhost:" + port + "/host/file/" + 2);
+
+        requestGet.setHeader("Authorization", "Bearer " + token);
+
+        HttpResponse responseGet = HttpClientBuilder.create().build().execute(requestGet);
+
+        HttpEntity entityGet = responseGet.getEntity();
+
+        String responseStringGet = EntityUtils.toString(entityGet, "UTF-8");
+
+        assertEquals(json, responseStringGet);
+        assertThat(responseGet.getStatusLine().getStatusCode(), equalTo(HttpStatus.SC_OK));
+    }
+
+    @Test
+    @Order(3)
+    public void testGetAllFileHosts() throws Exception {
+        ArrayList<HostFile> expectedListFile = new ArrayList<>();
+
+        HostFile host = new HostFile();
+        host.setId(2);
+        host.setMd5("randommd5value");
+        host.setFqHost("hostFq");
+        host.setHubId(1);
+        host.setHubFq("hubfq");
+
+        HostFile hub1 = new HostFile();
+        hub1.setId(1);
+        hub1.setFqHost("hubfq");
+        hub1.setMd5("hubmd5");
+        hub1.setHubId(1);
+        hub1.setHubFq("hubfq");
+
+        expectedListFile.add(hub1);
+        expectedListFile.add(host);
+        String expectedJson = gson.toJson(expectedListFile);
+
+        // Asserting get request
+        HttpGet requestGet = new HttpGet("http://localhost:" + port + "/host/file");
+
+        requestGet.setHeader("Authorization", "Bearer " + token);
+
+        HttpResponse responseGet = HttpClientBuilder.create().build().execute(requestGet);
+
+        HttpEntity entityGet = responseGet.getEntity();
+
+        String responseStringGet = EntityUtils.toString(entityGet, "UTF-8");
+
+        assertEquals(expectedJson, responseStringGet);
+        assertThat(responseGet.getStatusLine().getStatusCode(), equalTo(HttpStatus.SC_OK));
+
+    }
+
+    @Test
+    @Order(4)
+    public void testDeleteNonExistentFileHost() throws Exception {
+        HttpDelete delete = new HttpDelete("http://localhost:" + port + "/host/file/" + 99999);
+
+        // Header
+        delete.setHeader("Authorization", "Bearer " + token);
+
+        HttpResponse deleteResponse = HttpClientBuilder.create().build().execute(delete);
+
+        HttpEntity entityDelete = deleteResponse.getEntity();
+
+        String responseStringGet = EntityUtils.toString(entityDelete, "UTF-8");
+
+        // Parsin respponse as JSONObject
+        JSONObject responseAsJson = new JSONObject(responseStringGet);
+
+        // Creating string from Json that was given as a response
+        String actual = responseAsJson.get("message").toString();
+        // Creating expected message as JSON Object from the data that was sent towards endpoint
+        String expected = "Record does not exist";
+
+        assertEquals(expected, actual);
+        assertThat(deleteResponse.getStatusLine().getStatusCode(), equalTo(HttpStatus.SC_NOT_FOUND));
+    }
+
+    @Test
+    @Order(5)
+    public void testDeleteFileHost() throws Exception {
+        HttpDelete delete = new HttpDelete("http://localhost:" + port + "/host/file/" + 2);
+
+        // Header
+        delete.setHeader("Authorization", "Bearer " + token);
+
+        HttpResponse deleteResponse = HttpClientBuilder.create().build().execute(delete);
+
+        HttpEntity entityDelete = deleteResponse.getEntity();
+
+        String responseStringGet = EntityUtils.toString(entityDelete, "UTF-8");
+
+        // Parsin respponse as JSONObject
+        JSONObject responseAsJson = new JSONObject(responseStringGet);
+
+        // Creating string from Json that was given as a response
+        String actual = responseAsJson.get("message").toString();
+
+        // Creating expected message as JSON Object from the data that was sent towards endpoint
+        String expected = "Host deleted";
+
+        assertEquals(expected, actual);
+        assertThat(deleteResponse.getStatusLine().getStatusCode(), equalTo(HttpStatus.SC_OK));
+    }
+
+}
